paddleocr_service/requirements.txt
# PaddleOCR Service Dependencies
# Install in a SEPARATE virtual environment

flask>=3.0.0
paddlepaddle>=2.6.0
paddleocr>=2.9.0
paddlex>=3.0.0
pdf2image>=1.17.0
Pillow>=10.0.0
numpy>=1.24.0
